[Feature] Detail: Mapping path into json
