+++ pages/details/[detail].tsx
@@ -1,5 +1,5 @@
 import { DefaultLayout } from "@layouts/DefaultLayout";
-import { DetailsScreen } from "@screens/app/DetailsScreen";
+import { DetailsScreen } from "@screens/Details";
 
 const Details: IPageComponent = ({}) => {
   return <DetailsScreen />;
