+++ screens/Details/index.tsx
@@ -1,0 +1,59 @@
+import { ContentComponent } from "@components/Content";
+import { WidgetComponent } from "@components/Widget";
+import { Projects } from "@configs/detail";
+import Head from "next/head";
+
+export const DetailsScreen: IComponent = ({}) => {
+  const projectPrefix = window.location.pathname.split("/").pop();
+  const projectDetail = Projects[projectPrefix!];
+
+  /**
+   * Fallback to warning if project not found
+   */
+  if (!projectDetail) {
+    return (
+      <div
+        className="w-full flex justify-center items-center bg-white dark:bg-default"
+        style={{ height: "80vh" }}
+      >
+        <h1 className="text-2xl text-black dark:text-white opacity-80">
+          THIS PROJECT IS NOT <span className="font-bold">AVAILABLE</span>
+        </h1>
+      </div>
+    );
+  }
+
+  const { data } = projectDetail;
+
+  const content = data.content;
+  const widget = data.widget_detail_1;
+  const widget2 = data.widget_detail_2;
+
+  return (
+    <div className="">
+      <Head>
+        <title>Create Next App</title>
+        <meta name="description" content="Generated by create next app" />
+        <link rel="icon" href="/favicon.ico" />
+      </Head>
+      <main className="details min-h-screen ">
+        <div className="banner py-48 bg-[#f7f7f7]">
+          <div className="flex items-center justify-center ">
+            <h1 className="text-6xl font-bold ">OUR WORKS</h1>
+          </div>
+        </div>
+        <div className="container mx-auto py-40">
+          <div className="grid grid-rows-3 xl:grid-rows-none xl:grid-cols-3 gap-10">
+            <div className="row-span-2 xl:col-span-2 flex flex-col p-2">
+              <ContentComponent {...content} />
+            </div>
+            <div className="flex flex-col justify-between  xl:flex xl:flex-col p-2">
+              <WidgetComponent {...widget} />
+              <WidgetComponent {...widget2} />
+            </div>
+          </div>
+        </div>
+      </main>
+    </div>
+  );
+};
