--- screens/app/DetailsScreen/index.tsx
@@ -1,75 +1,0 @@
-import { ContentComponent } from "@components/Content";
-import { WidgetComponent } from "@components/Widget";
-import Head from "next/head";
-
-export const DetailsScreen: IComponent = ({}) => {
-  const content = {
-    imageSrc: "abcxaudfasdfsfsd",
-    heading: "DMS",
-    subheading: "Backend server",
-    paragraphs: [
-      "Definition: Business strategy can be understood as the course of action or set of  decisions which assist the entrepreneurs in achieving specific business objectives.",
-      "Definition: Business strategy can be understood as the course of action or set of  decisions which assist the entrepreneurs in achieving specific business objectives.",
-      "Definition: Business strategy can be understood as the course of action or set of  decisions which assist the entrepreneurs in achieving specific business objectives.",
-    ],
-    techs: [
-      "Golang",
-      "Gin",
-      "Gorm",
-      "MSSQL",
-      "Docker",
-      "CI/CD github action",
-      "Swagger documentation",
-      "Paho MQTT client",
-      "EMQX MQTT broker",
-    ],
-    scope:
-      "Create HTTP API for client apps to control doors in university. Set up communication between BE server with Raspberry gateway via MQTT protocol/APIs.Store/retrieve data, logs and schedulers of doors/gateways.",
-  };
-  const widget = {
-    title: "Project information",
-    project: {
-      date: "12/1/20021",
-      location: "HCM",
-      client: "string",
-      category: "UX",
-      link: "https://nextjs.org",
-    },
-  };
-  const widget2 = {
-    title: "contact information",
-    contact: {
-      address: "HCM",
-      mail: "@gmial.com",
-      phone: "034982343",
-    },
-  };
-
-  return (
-    <div className="">
-      <Head>
-        <title>Create Next App</title>
-        <meta name="description" content="Generated by create next app" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-      <main className="details min-h-screen ">
-        <div className="banner py-48 bg-[#f7f7f7]">
-          <div className="flex items-center justify-center ">
-            <h1 className="text-6xl font-bold ">OUR WORKS</h1>
-          </div>
-        </div>
-        <div className="container mx-auto py-40">
-          <div className="grid grid-rows-3 xl:grid-rows-none xl:grid-cols-3 gap-10">
-            <div className="row-span-2 xl:col-span-2 flex flex-col p-2">
-              <ContentComponent {...content} />
-            </div>
-            <div className="flex flex-col justify-between  xl:flex xl:flex-col p-2">
-              <WidgetComponent {...widget} />
-              <WidgetComponent {...widget2} />
-            </div>
-          </div>
-        </div>
-      </main>
-    </div>
-  );
-};
